(tui) readline widget + unicode helpers
This adds the first real round of the readline widget. It covers some of
the basic / expected navigations, completion and so on, but clearly a
lot of things remain for it to be viable versus libreadline style work.

Since this also is heavier on conversion back and forth between
representations, basic ucs4-utf8 conversion helpers are also added to
the public interface.