(tui) ucs4utf8_s and cli_command handler
This adds a helper around ucs4utf8 that adds null termination.
Introduce a new entry in the handler table, 'cli_command'.

This one is a way of letting each tui window provide integration with an
outer shell command-line input, as alternative or complement to a
readline widget.

The specific shmif-path is still undecided. It is possible that this
will simply be MESSAGE as that is not used for anything else here at the
moment. Other candidates are COREOPT, though it is probably better to
keep that for its purpose of offline key/value config sture in the
spirit of libretro from whence it came.

The big test case for this is to see how it integrates with the menu hud
in durden, i.e. map /target/cli to access it, and then have different
update/forward hooks for repopulating the completion set and indicating
error message.