(wayland) fix edge case sending configure for (0,0)
This triggered a bug in some clients, though the spec is a bit iffy
here on what is the proper action. Initial config was sent based on
whatever we wanted, but updates that only targeted selection or wnd
state had w,h set to 0, allowing the client to 'pick', but instead
of keeping the current size, some clients just revert to an initial
default. Fix by sending the last sent hint or current acon size.